Move DRPCRecorder into the recorder package
