+++ coderd/aibridged/pool.go
@@ -19,6 +19,7 @@ import (
 	"github.com/coder/coder/v2/aibridge"
 	"github.com/coder/coder/v2/aibridge/keypool"
 	"github.com/coder/coder/v2/aibridge/mcp"
+	"github.com/coder/coder/v2/aibridge/recorder"
 	"github.com/coder/coder/v2/aibridge/tracing"
 	"github.com/coder/quartz"
 )
@@ -228,7 +229,7 @@ func (p *CachedBridgePool) Acquire(ctx context.Context, req Request, clientFn Cl
 
 	span.AddEvent("cache_miss")
 	providerVersion := p.providerVersion.Load()
-	recorder := aibridge.NewRecorder(
+	rec := aibridge.NewRecorder(
 		p.logger.Named("recorder"),
 		p.tracer,
 		func(clientCtx context.Context) (aibridge.Recorder, error) {
@@ -239,7 +240,7 @@ func (p *CachedBridgePool) Acquire(ctx context.Context, req Request, clientFn Cl
 				return nil, xerrors.Errorf("acquire client: %w", err)
 			}
 
-			return &DRPCRecorder{apiKeyID: req.APIKeyID, client: client}, nil
+			return recorder.NewDRPCRecorder(req.APIKeyID, client), nil
 		},
 	)
 
@@ -266,7 +267,16 @@ func (p *CachedBridgePool) Acquire(ctx context.Context, req Request, clientFn Cl
 			}
 		}
 
-		bridge, err := aibridge.NewRequestBridge(ctx, p.loadProviders(), recorder, mcpServers, p.logger, p.metrics, p.tracer, aibridge.WithClock(p.clock))
+		bridge, err := aibridge.NewRequestBridge(
+			ctx,
+			p.loadProviders(),
+			rec,
+			mcpServers,
+			p.logger,
+			p.metrics,
+			p.tracer,
+			aibridge.WithClock(p.clock),
+		)
 		if err != nil {
 			return nil, xerrors.Errorf("create new request bridge: %w", err)
 		}
